requirements.txt
requests>=2.28.0
notion-client==2.2.1
python-dotenv>=1.0.0
